+++ cli/src/commands/workspace/forget.rs
@@ -15,6 +15,8 @@
 use clap_complete::ArgValueCandidates;
 use itertools::Itertools as _;
 use jj_lib::ref_name::WorkspaceNameBuf;
+use jj_lib::workspace_store::SimpleWorkspaceStore;
+use jj_lib::workspace_store::WorkspaceStore as _;
 use tracing::instrument;
 
 use crate::cli_util::CommandHelper;
@@ -42,6 +44,7 @@ pub fn cmd_workspace_forget(
     args: &WorkspaceForgetArgs,
 ) -> Result<(), CommandError> {
     let mut workspace_command = command.workspace_helper(ui)?;
+    let repo_path = workspace_command.repo_path().to_path_buf();
 
     let wss = if args.workspaces.is_empty() {
         vec![workspace_command.workspace_name().to_owned()]
@@ -63,8 +66,19 @@ pub fn cmd_workspace_forget(
     // bundle every workspace forget into a single transaction, so that e.g.
     // undo correctly restores all of them at once.
     let mut tx = workspace_command.start_transaction();
-    wss.iter()
-        .try_for_each(|ws| tx.repo_mut().remove_wc_commit(ws))?;
+
+    let workspace_store = SimpleWorkspaceStore::load(&repo_path)?;
+
+    for ws in &wss {
+        tx.repo_mut().remove_wc_commit(ws)?;
+
+        // This is to make sure not to throw error for workspaces created before
+        // this change.
+        if workspace_store.exists(ws) {
+            workspace_store.remove_path(ws)?;
+        }
+    }
+
     let description = if let [ws] = wss.as_slice() {
         format!("forget workspace {}", ws.as_symbol())
     } else {