+++ cli/src/commands/workspace/root.rs
@@ -14,26 +14,57 @@
 
 use std::io::Write as _;
 
+use clap_complete::ArgValueCandidates;
 use jj_lib::file_util;
+use jj_lib::ref_name::WorkspaceNameBuf;
+use jj_lib::workspace_store::SimpleWorkspaceStore;
+use jj_lib::workspace_store::WorkspaceStore as _;
 use tracing::instrument;
 
 use crate::cli_util::CommandHelper;
 use crate::command_error::CommandError;
 use crate::command_error::user_error;
+use crate::complete;
 use crate::ui::Ui;
 
-/// Show the current workspace root directory
+/// Show the workspace root directory
 #[derive(clap::Args, Clone, Debug)]
-pub struct WorkspaceRootArgs {}
+pub struct WorkspaceRootArgs {
+    /// Name of the workspace (defaults to the current)
+    #[arg(long, short, value_name = "WORKSPACE", add = ArgValueCandidates::new(complete::workspaces))]
+    workspace: Option<WorkspaceNameBuf>,
+}
 
 #[instrument(skip_all)]
 pub fn cmd_workspace_root(
     ui: &mut Ui,
     command: &CommandHelper,
-    _args: &WorkspaceRootArgs,
+    args: &WorkspaceRootArgs,
 ) -> Result<(), CommandError> {
-    let loader = command.workspace_loader()?;
-    let path_bytes = file_util::path_to_bytes(loader.workspace_root()).map_err(user_error)?;
+    let workspace_command = command.workspace_helper(ui)?;
+    let repo_path = workspace_command.repo_path().to_path_buf();
+
+    let name = if let Some(ws_name) = &args.workspace {
+        ws_name
+    } else {
+        workspace_command.workspace_name()
+    };
+
+    let workspace_store = SimpleWorkspaceStore::load(&repo_path)?;
+
+    let path = if workspace_store.exists(&name.to_owned()) {
+        let workspace_proto = workspace_store.get_path(&name.to_owned())?;
+        dunce::canonicalize(workspace_proto.path)?
+    } else if args.workspace.is_some() {
+        return Err(user_error(format!(
+            "No such workspace: {}",
+            name.as_symbol()
+        )));
+    } else {
+        workspace_command.workspace_root().to_path_buf()
+    };
+
+    let path_bytes = file_util::path_to_bytes(&path).map_err(user_error)?;
     ui.stdout().write_all(path_bytes)?;
     writeln!(ui.stdout())?;
     Ok(())